Port

+++ src/engine/compute.d
@@ -20,7 +20,7 @@ import validation : pushLabel, popLabel, nameVulkanObject;
 /** Compute structure with shaders, command buffer and pipelines
  */
 struct Compute {
-  uint lastTick;
+  size_t lastTick;
   ParticleSystem system;
   Shader[] shaders;                           /// Compute shader objects
   VkCommandBuffer[][string] commands;   /// Command buffers
@@ -102,7 +102,7 @@ void updateComputeUBO(ref App app, uint syncIndex = 0){
     auto shader = app.compute.shaders[s];
     for(uint d = 0; d < shader.descriptors.length; d++) {
       if(shader.descriptors[d].type == VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER){
-        uint now = SDL_GetTicks();
+        size_t now = SDL_GetTicks();
         ParticleUniformBuffer buffer = {
           position: app.compute.system.position.xyzw,
           gravity: app.compute.system.gravity.xyzw,
